+++ crates/forge_api/Cargo.toml
@@ -9,7 +9,7 @@ async-trait.workspace = true
 forge_domain.workspace = true
 forge_stream.workspace = true
 forge_services.workspace = true
-forge_walker.workspace = true
+
 forge_infra.workspace = true
 forge_snaps.workspace = true
 serde_yml.workspace = true
