+++ crates/forge_main/Cargo.toml
@@ -13,7 +13,7 @@ clap.workspace = true
 console.workspace = true
 forge_api.workspace = true
 forge_domain.workspace = true
-forge_walker.workspace = true
+forge_infra.workspace = true
 forge_display.workspace = true
 forge_tracker.workspace = true
 forge_snaps.workspace = true
