+++ crates/forge_snaps/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 [dependencies]
 anyhow.workspace = true
 forge_fs.workspace = true
-forge_walker.workspace = true
+
 fnv_rs.workspace = true
 base64.workspace = true
 serde_json.workspace = true
