--- crates/forge_walker/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "forge_walker"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-ignore.workspace = true
-tokio.workspace = true
-anyhow.workspace = true
-derive_setters.workspace = true
-
-[dev-dependencies]
-pretty_assertions.workspace = true
-tempfile.workspace = true
