--- crates/forge_walker/src/lib.rs
@@ -1,3 +1,0 @@
-mod walker;
-
-pub use walker::{File, Walker};
